Revert "Removed login, improved pipeline, added eslint"
This reverts commit b8f40d4b5e6f36a0c2c90787d18ffa77c1693691.

+++ _scripts/ci-dep.sh
@@ -1,19 +1,27 @@
 #!/bin/bash
 set -x
+echo $TRAVIS_BRANCH
 
-# encrypt key
-openssl aes-256-cbc -K $encrypted_e9ce4c9f2f9f_key -iv $encrypted_e9ce4c9f2f9f_iv -in deploy-key.enc -out deploy-key -d
-rm deploy-key.enc
-chmod 600 deploy-key
-mv deploy-key ~/.ssh/id_rsa
+# Test if branch is master branch
+if [ $TRAVIS_BRANCH == "master" ]; then
 
-# change to build directory, initialize local repository and push to server
-cd build
-git init
+    # encrypt key
+    openssl aes-256-cbc -K $encrypted_e9ce4c9f2f9f_key -iv $encrypted_e9ce4c9f2f9f_iv -in deploy-key.enc -out deploy-key -d
+    rm deploy-key.enc
+    chmod 600 deploy-key
+    mv deploy-key ~/.ssh/id_rsa
 
-git remote add deploy "$SERVER_USER@$SERVER_IP:$SERVER_HTML_PATH"
-git config user.name = "Travis CI"
+    # change to build directory, initialize local repository and push to server
+    cd build
+    git init
 
-git add .
-git commit -m "Deploy Commit: $TRAVIS_COMMIT"
-git push --force deploy master
+    git remote add deploy "$SERVER_USER@$SERVER_IP:$SERVER_HTML_PATH"
+    git config user.name = "Travis CI"
+
+    git add .
+    git commit -m "Deploy Commit: $TRAVIS_COMMIT"
+    git push --force deploy master
+
+else
+    echo "Not on master branch, not deploying"
+fi
